Apply feedback about right annotations
Provide the right annotations
Move UriJavaAccessor to impl package

Adapt mima filters
