+++ akka-http-core/src/main/java/akka/http/javadsl/model/Uri.java
@@ -7,7 +7,7 @@ package akka.http.javadsl.model;
 import java.nio.charset.Charset;
 
 import akka.http.impl.model.JavaUri;
-import akka.http.scaladsl.model.UriJavaAccessor;
+import akka.http.impl.model.UriJavaAccessor;
 import akka.parboiled2.ParserInput$;
 
 import java.util.Optional;
