+++ akka-http-core/src/main/scala/akka/http/impl/model/JavaQuery.scala
@@ -7,7 +7,7 @@ package akka.http.impl.model
 import java.util.Optional
 import java.{ util ⇒ ju }
 
-import akka.annotation.DoNotInherit
+import akka.annotation.InternalApi
 import akka.http.impl.model.parser.CharacterClasses
 import akka.http.impl.util.StringRendering
 import akka.http.javadsl.model.HttpCharset
@@ -21,8 +21,8 @@ import scala.collection.JavaConverters._
 import akka.http.impl.util.JavaMapping.Implicits._
 
 /** INTERNAL API */
-@DoNotInherit
-case class JavaQuery(query: sm.Uri.Query) extends jm.Query {
+@InternalApi
+final case class JavaQuery(query: sm.Uri.Query) extends jm.Query {
   override def get(key: String): Optional[String] = query.get(key).asJava
   override def toMap: ju.Map[String, String] = query.toMap.asJava
   override def toList: ju.List[Pair[String, String]] = query.map(_.asJava).asJava
