+++ akka-http-core/src/main/scala/akka/http/impl/model/JavaUri.scala
@@ -8,14 +8,14 @@ import java.nio.charset.Charset
 import java.util.Optional
 import java.{ lang ⇒ jl }
 
-import akka.annotation.DoNotInherit
+import akka.annotation.InternalApi
 import akka.http.scaladsl.model.Uri.ParsingMode
 import akka.http.javadsl.{ model ⇒ jm }
 import akka.http.scaladsl.{ model ⇒ sm }
 import akka.http.impl.util.JavaMapping.Implicits._
 
 /** INTERNAL API */
-@DoNotInherit
+@InternalApi
 case class JavaUri(uri: sm.Uri) extends jm.Uri {
   def isRelative: Boolean = uri.isRelative
   def isAbsolute: Boolean = uri.isAbsolute
