+++ akka-http-core/src/main/scala/akka/http/javadsl/settings/SettingsCompanion.scala
@@ -1,11 +1,8 @@
 package akka.http.javadsl.settings
 
 import akka.actor.ActorSystem
-import akka.annotation.InternalApi
 import com.typesafe.config.Config
 
-/** INTERNAL API */
-@InternalApi
 trait SettingsCompanion[T] {
 
   /**
