+++ akka-http-tests/src/test/scala/akka/http/scaladsl/server/IntegrationRoutingSpec.scala
@@ -4,7 +4,6 @@
 package akka.http.scaladsl.server
 
 import akka.actor.ActorSystem
-import akka.annotation.InternalApi
 import akka.http.scaladsl.{ Http, TestUtils }
 import akka.http.scaladsl.client.RequestBuilding
 import akka.http.scaladsl.model.{ HttpRequest, HttpResponse }
@@ -13,8 +12,7 @@ import akka.testkit.{ AkkaSpec, TestKit }
 import org.scalatest.concurrent.{ IntegrationPatience, ScalaFutures }
 import org.scalatest.{ BeforeAndAfterAll, Matchers, WordSpecLike }
 
-/** INTERNAL API - not (yet?) ready for public consuption */
-@InternalApi
+/** not (yet?) ready for public consuption */
 private[akka] trait IntegrationRoutingSpec extends WordSpecLike with Matchers with BeforeAndAfterAll
   with Directives with RequestBuilding
   with ScalaFutures with IntegrationPatience {
