+++ akka-http2-support/src/main/scala/akka/http/impl/engine/http2/Http2Protocol.scala
@@ -15,7 +15,7 @@ import akka.util.ByteString
  * https://tools.ietf.org/html/rfc7540
  */
 @InternalApi
-object Http2Protocol {
+private[http2] object Http2Protocol {
   // constants defined in the spec
 
   /**
